Refactor Loop and Synchronizer related code.
