+++ game/src/main/kotlin/xlitekt/game/loop/sync/WorldLoginSynchronizer.kt
@@ -1,0 +1,10 @@
+package xlitekt.game.loop.sync
+
+/**
+ * @author Jordan Abraham
+ */
+class WorldLoginSynchronizer : Synchronizer() {
+    override fun run() {
+        world.processLoginRequests()
+    }
+}
