+++ game/src/main/kotlin/xlitekt/game/loop/sync/WorldLogoutSynchronizer.kt
@@ -1,0 +1,10 @@
+package xlitekt.game.loop.sync
+
+/**
+ * @author Jordan Abraham
+ */
+class WorldLogoutSynchronizer : Synchronizer() {
+    override fun run() {
+        world.processLogoutRequests()
+    }
+}
